(PE-17701) Separate update check from telemetry
This deprecates `check-for-updates!` and replaces it with two new
functions, `send-telemetry` and `check-for-update`, to allow sending
telemetry data without reporting on version updates, and check for
version updates without sending any telemetry data. The
`check-for-update` function filters out any parameters that are
unnecessary for a version check. The new functions are synchronous. If
you want the behavior of the old function just wrap them in futures.

This also attempts to improve the error handling and logging.